fix: 修正 ClaimedDelayedTask 中 referenceIdAsLong 方法的返回类型
- 将 referenceIdAsLong 方法的返回类型从 int 修改为 long，以确保正确处理业务主键场景中的长整型值。

+++ mahu-task/src/main/java/cool/houge/mahu/task/handler/ClaimedDelayedTask.java
@@ -33,7 +33,7 @@ public record ClaimedDelayedTask(
     }
 
     /// 将 referenceId 转换为 long，通常用于业务主键场景。
-    public int referenceIdAsLong() {
+    public long referenceIdAsLong() {
         return Long.parseLong(referenceId);
     }
